Differentiate effects, potions, and splashes.
This will allow players to drink certain potions while still
being restricted from throwing them. For example, a player
can drink a potion of harming (if they really wanted to!),
yet be unable to throw it at an enemy.
The code will be updated to take advantage of this soon.

"deny-effects" can be used as a global restriction for both
drinks and splashes. It can also deny beacon effects.

The code will be updated to take advantage of this soon.