getEffectString() => getFriendlyEffectNames()
getFriendlyEffectNames() takes a String array with effect names
(ideally from a command) as an argument, and returns a List<String>
with all of the effect names that would be used in the configuration
file. It accounts for some alternate effect names.
If there were no matches, then it returns an empty List.

Also renovated InitArrays() to make "positive", "negative", and
"all" ArrayLists for use as shortcutting.